+++ integration_test/cases/repo.exs
@@ -166,21 +166,14 @@ defmodule Ecto.Integration.RepoTest do
     assert validate_unique(on_update, :title, on: TestRepo).errors == []
   end
 
-  test "validate_unique/3 scope" do
+  test "validate_unique/3 with scope" do
     import Ecto.Changeset
     post = TestRepo.insert(%Post{title: "hello", text: "world"})
 
     on_insert = cast(%Post{}, %{"title" => "hello", "text" => "elixir"}, ~w(title), ~w(text))
     assert validate_unique(on_insert, :title, on: TestRepo).errors == [title: "has already been taken"]
     assert validate_unique(on_insert, :title, scope: [:text], on: TestRepo).errors == []
-  end
 
-  test "validate_unique/3 scope with a non-existent field will raise Ecto.QueryError" do
-    import Ecto.Changeset
-    post = TestRepo.insert(%Post{title: "hello", text: "world"})
-
-    on_insert = cast(%Post{}, %{"title" => "hello"}, ~w(title), ~w(text))
-    assert validate_unique(on_insert, :title, on: TestRepo).errors == [title: "has already been taken"]
     assert_raise(Ecto.QueryError, fn ->
       validate_unique(on_insert, :title, scope: [:non_existent], on: TestRepo).errors == []
     end)